+++ app/assets/themes/degrowth/partials/_input.scss
@@ -4,3 +4,23 @@
 input:checked ~ .switch-paddle {
   background-color: $success;
 }
+
+[type="text"],
+[type="password"],
+[type="date"],
+[type="datetime"],
+[type="datetime-local"],
+[type="month"],
+[type="week"],
+[type="email"],
+[type="number"],
+[type="search"],
+[type="tel"],
+[type="time"],
+[type="url"],
+[type="color"],
+textarea {
+  background-color: $white-transparent;
+  border: $border-thin;
+  border-radius: 0;
+}
