+++ app/cells/decidim/content_blocks/highlighted_content_banner/show.erb
@@ -1,5 +1,5 @@
 <section class="extended highligted-content-banner home-section"
-          style="background-image:url(<%= asset_path(instance_prefix_path("slider.jpg")) %>);">
+          style="background-image:url(<%= asset_path(current_organization.instance_prefix_path("slider.jpg")) %>);">
     <div class="black-background">
       <span>
         <%= decidim_sanitize translated_attribute current_organization.highlighted_content_banner_short_description %>
