+++ app/views/layouts/decidim/_email_banner.html.erb
@@ -11,7 +11,7 @@
             <% end %>
 
             <%= link_to url, target: :blank do %>
-              <%= image_tag asset_url(instance_prefix_path("banner-email.jpg"), host: url), alt: organization.name %>
+              <%= image_tag asset_url(@organization.instance_prefix_path("banner-email.jpg"), host: url), alt: organization.name %>
             <% end %>
           </td>
         </tr>
