+++ config/initializers/instance_helper.rb
@@ -1,6 +1,27 @@
 # frozen_string_literal: true
 
-Rails.application.config.to_prepare do
-  Decidim::ApplicationController.helper(InstanceHelper)
-  Decidim::ViewModel.include(InstanceHelper)
+Decidim::Organization.class_eval do
+  def instance_name
+    if citiesforchange?
+      :citiesforchange
+    elsif degrowth?
+      :degrowth
+    end
+  end
+
+  def instance_prefix_path(path)
+    if citiesforchange?
+      path
+    else
+      "#{instance_name}/#{path}"
+    end
+  end
+
+  def citiesforchange?
+    host.match?("citiesforchange")
+  end
+
+  def degrowth?
+    host.match?("degrowth") || host.match?("localhost") # DEBUG
+  end
 end
